move ChoiceInput to Widget folder
