+++ Composer/packages/extensions/visual-designer/__tests__/widgets/ChoiceInput.test.tsx
@@ -4,8 +4,8 @@
 import React from 'react';
 import { render, cleanup } from 'react-testing-library';
 
-import { ChoiceInput } from '../../../src/components/nodes/steps/ChoiceInput';
-import { ObiTypes } from '../../../src/constants/ObiTypes';
+import { ChoiceInput } from '../../src/widgets/ChoiceInput';
+import { ObiTypes } from '../../src/constants/ObiTypes';
 
 describe('<ChoiceInput />', () => {
   let renderResult, id, data, onEvent, onResize;
